Configure device by common function for all samples
