+++ examples/classification/staged_quantization_worker.py
@@ -30,14 +30,14 @@ from examples.common.example_logger import logger
 from examples.common.execution import ExecutionMode, prepare_model_for_execution
 from examples.common.model_loader import load_model
 from examples.common.utils import configure_logging, print_args, make_additional_checkpoints, get_name, \
-    print_statistics, is_pretrained_model_requested, log_common_mlflow_params, SafeMLFLow
+    print_statistics, is_pretrained_model_requested, log_common_mlflow_params, SafeMLFLow, configure_device
 from nncf.binarization.algo import BinarizationController
 from nncf.compression_method_api import CompressionLevel
 from nncf.initialization import register_default_init_args, default_criterion_fn
 from nncf.model_creation import create_compressed_model
 from nncf.quantization.algo import QuantizationController
 from nncf.utils import is_main_process
-from examples.classification.common import configure_device, set_seed, load_resuming_checkpoint
+from examples.classification.common import set_seed, load_resuming_checkpoint
 
 
 class KDLossCalculator:
