+++ examples/common/utils.py
@@ -27,6 +27,8 @@ from typing import Tuple
 from PIL import Image
 import torch.utils.data as data
 
+from examples.common.distributed import configure_distributed
+from examples.common.execution import ExecutionMode, get_device
 from examples.common.sample_config import SampleConfig
 from torch.utils.tensorboard import SummaryWriter
 from texttable import Texttable
@@ -138,6 +140,15 @@ class SafeMLFLow:
         if self._is_enabled():
             return mlflow
         return None
+
+
+def configure_device(current_gpu, config: SampleConfig):
+    config.current_gpu = current_gpu
+    config.distributed = config.execution_mode in (ExecutionMode.DISTRIBUTED, ExecutionMode.MULTIPROCESSING_DISTRIBUTED)
+    if config.distributed:
+        configure_distributed(config)
+
+    config.device = get_device(config)
 
 
 def configure_logging(sample_logger, config):
